L23

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
@@ -15,11 +16,15 @@ class CategoryController extends Controller
 
     public function create()
     {
+        Gate::authorize('create', Category::class);
+
         return view('category.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create', Category::class);
+
         $validated = $request->validate([
             'name' => 'required|unique:categories,name',
             'color' => 'nullable',
@@ -32,6 +37,8 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        Gate::authorize('destroy', $category);
+
         $category->articles()->detach();
         $category->delete();
